+++ tests/functional/test_cli.py
@@ -33,7 +33,7 @@ def test_run_no_paths(runner, monkeypatch, tmpdir):
 
     result = runner.invoke(cli.main, [])
 
-    assert "actual: 50.0%" in result.output
+    assert "actual: 51.4%" in result.output
     assert 1 == result.exit_code
 
 
@@ -41,39 +41,39 @@ def test_run_no_paths(runner, monkeypatch, tmpdir):
     "flags,exp_result,exp_exit_code",
     (
         # no flags
-        ([], 50.0, 1),
+        ([], 51.4, 1),
         # ignore init module
-        (["-I"], 50.0, 1),
+        (["-I"], 51.5, 1),
         # ignore module docs
-        (["-M"], 50.0, 1),
+        (["-M"], 51.6, 1),
         # ignore semiprivate docs
-        (["-s"], 50.8, 1),
+        (["-s"], 52.4, 1),
         # ignore private docs
-        (["-p"], 51.6, 1),
-        # ignore property getter/setter decorators
+        (["-p"], 53.1, 1),
+        # ignore property getter/setter/deleter decorators
         (["-P"], 50.0, 1),
         # ignore property setter decorators
-        (["-S"], 50.0, 1),
+        (["-S"], 51.5, 1),
         # ignore magic method docs
-        (["-m"], 50.0, 1),
+        (["-m"], 51.5, 1),
         # ignore init method docs
-        (["-i"], 49.2, 1),
+        (["-i"], 50.7, 1),
         # ignore nested funcs
-        (["-n"], 49.2, 1),
+        (["-n"], 50.7, 1),
         # ignore nested classes
-        (["-C"], 50.8, 1),
+        (["-C"], 52.2, 1),
         # ignore @typing.overload-decorated functions
-        (["-O"], 50.0, 1),
+        (["-O"], 51.6, 1),
         # ignore regex
-        (["-r", "^get$"], 50.0, 1),
+        (["-r", "^get$"], 51.5, 1),
         # whitelist regex
         (["-w", "^get$"], 50.0, 1),
         # exclude file
-        (["-e", os.path.join(SAMPLE_DIR, "partial.py")], 62.5, 1),
+        (["-e", os.path.join(SAMPLE_DIR, "partial.py")], 63.4, 1),
         # exclude file which doesn't exist
-        (["-e", os.path.join(SAMPLE_DIR, "does.not.exist")], 50.0, 1),
+        (["-e", os.path.join(SAMPLE_DIR, "does.not.exist")], 51.4, 1),
         # fail under
-        (["-f", "40"], 50.0, 0),
+        (["-f", "40"], 51.4, 0),
     ),
 )
 def test_run_shortflags(flags, exp_result, exp_exit_code, runner):
@@ -89,21 +89,21 @@ def test_run_shortflags(flags, exp_result, exp_exit_code, runner):
 @pytest.mark.parametrize(
     "flags,exp_result,exp_exit_code",
     (
-        (["--ignore-init-module"], 50.0, 1),
-        (["--ignore-module"], 50.0, 1),
-        (["--ignore-semiprivate"], 50.8, 1),
-        (["--ignore-private"], 51.6, 1),
+        (["--ignore-init-module"], 51.5, 1),
+        (["--ignore-module"], 51.6, 1),
+        (["--ignore-semiprivate"], 52.4, 1),
+        (["--ignore-private"], 53.1, 1),
         (["--ignore-property-decorators"], 50.0, 1),
-        (["--ignore-setters"], 50.0, 1),
-        (["--ignore-magic"], 50.0, 1),
-        (["--ignore-init-method"], 49.2, 1),
-        (["--ignore-nested-functions"], 49.2, 1),
-        (["--ignore-nested-classes"], 50.8, 1),
-        (["--ignore-overloaded-functions"], 50.0, 1),
-        (["--ignore-regex", "^get$"], 50.0, 1),
+        (["--ignore-setters"], 51.5, 1),
+        (["--ignore-magic"], 51.5, 1),
+        (["--ignore-init-method"], 50.7, 1),
+        (["--ignore-nested-functions"], 50.7, 1),
+        (["--ignore-nested-classes"], 52.2, 1),
+        (["--ignore-overloaded-functions"], 51.6, 1),
+        (["--ignore-regex", "^get$"], 51.5, 1),
         (["--whitelist-regex", "^get$"], 50.0, 1),
-        (["--exclude", os.path.join(SAMPLE_DIR, "partial.py")], 62.5, 1),
-        (["--fail-under", "40"], 50.0, 0),
+        (["--exclude", os.path.join(SAMPLE_DIR, "partial.py")], 63.4, 1),
+        (["--fail-under", "40"], 51.4, 0),
     ),
 )
 def test_run_longflags(flags, exp_result, exp_exit_code, runner):
@@ -119,9 +119,9 @@ def test_run_longflags(flags, exp_result, exp_exit_code, runner):
 @pytest.mark.parametrize(
     "flags,exp_result,exp_exit_code",
     (
-        (["-i", "-I", "-r" "^method_foo$"], 50.0, 1),
-        (["-s", "-p", "-M"], 53.1, 1),
-        (["-m", "-f", "45"], 50.0, 0),
+        (["-i", "-I", "-r" "^method_foo$"], 51.6, 1),
+        (["-s", "-p", "-M"], 54.9, 1),
+        (["-m", "-f", "45"], 51.5, 0),
     ),
 )
 def test_run_multiple_flags(flags, exp_result, exp_exit_code, runner):